requirements.txt
torch
torchvision
numpy
pandas
matplotlib
scikit-learn
Pillow
streamlit
ollama